data: adding jobspec associated files
These files were programatically found by cloning each
respository, parsing the tokens in the jobspec file,
and then if a path existed relative to the repository,
it was added here and saved in the lookup for configs. We
need the lookup because often scripts/configs are shared
between jobspec files. A manifest is added for a mapping.

Signed-off-by: vsoch <vsoch@users.noreply.github.com>